Refactor: move versionGE into app_manager_utils
Something I found that can be improved, so release.js doesn't have to import a big app_manager.js...but this is not really related to the change in this PR...
